fix(camtrack): fix hook leak and update for 0.8.x+
Previously, each time the macro was run, it would create a new hook,
each of which would run on token movement. Now, the hook ID is tracked,
and the old hook removed if the macro is re-run.

Additionally, the 0.8.x Foundry update changed the token hook API, so
the hook arguments have been adjusted to reflect this.